pyproject.toml
[tool.poetry]
name = "SynThesis"
version = "0.1.0"
description = "A toolkit for generating and evaluating synthetic data in terms of utility, privacy, and similarity"
authors = ["Maryam Mohebi <maryam@mohebi.me"]
readme = "README.md"
packages = [{ include = "synthetic_data" }]

[tool.poetry.dependencies]
python = "3.10.10"
pomegranate = "0.15.0"
sdv = "^1.15.0"
sdmetrics = "^0.15.0"
jupyter = "^1.0.0"
ipywidgets = "^8.1.3"
ipykernel = "^6.29.5"
numpy = "^1.23.0"
autogluon-tabular = { version = "^1.1.1", extras = ["all"] }
pandas = "^2.2.2"
matplotlib = "^3.9.1"
scipy = "^1.12.0"
scikit-learn = "^1.4.0"
copulas = "0.11.0"
torch = "2.3.1"
torchvision = "0.18.1"
seaborn = "^0.13.2"
shap = "^0.46.0"
pyspark = "^3.5.1"
tqdm = "^4.66.4"
tensorflow = "^2.16.1"
tensorflow-probability = "^0.24.0"
arfpy = "^0.1.1"
anonymeter = "^1.0.0"
requests = ">=2.30,<2.31"
pmlb = "1.0.*"
ipython-autotime = "^0.3.2"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
mypy = "^1.7.0"
ruff = "^0.1.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.10"
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.black]
line-length = 110

[tool.ruff]
target-version = "py39"
line-length = 120
fix = true
select = ["ALL"]
ignore = ["D100", "D104", "ISC003", "TD003"]

[tool.ruff.pydocstyle]
convention = "google"